Attempted fix to support HTTPS proxying
My attempts to enable the etm-esdl app over HTTPS haven't worked, with
the app showing "No API definition provided.".

I think this is due to the app expecting HTTP only (that's how it's
running in the Docker container). I found this fix which, which I'm
hopeful will resolve the issue:

https://github.com/noirbizarre/flask-restplus/issues/565#issuecomment-562610603